# Change so Garbage Tensors don't use extra string stream to generate garbage data