Before this PR, WriteDataToBinding was still called even though the intent was to use garbage data. This change makes it so that arbitrary memory is created for the binding. The memory is memset to 0 so that garbage data results are deterministic and consistent.

**There is about a 2x performance increase with evaluate with this change since there is less time between GPU usage. Before this change, the GPU powers down / power up between iterations.**

I ran Squeezenet 300 iterations specifying GPU as device and compared the captured average performance.

Before this PR change:

  Load: 21.470000 ms
  Bind: 0.473307 ms
  **Evaluate: 18.5709 ms**
  Total Time: 40.5142 ms
  Working Set Memory usage (evaluate): 0 MB
  Dedicated Memory Usage (evaluate): 0 MB
  Shared Memory Usage (evaluate): 0 MB

After this PR change:

  Load: 46.911000 ms
  Bind: 0.685665 ms
  **Evaluate: 8.89519 ms**
  Total Time: 56.4919 ms
  Working Set Memory usage (evaluate): 0.00375366 MB
  Dedicated Memory Usage (evaluate): 0 MB
  Shared Memory Usage (evaluate): 0 MB